efoodapiv1/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.7
Django==4.2.13
django-ckeditor==6.7.1
django-geojson==4.1.0
django-js-asset==2.2.0
django-leaflet==0.30.0
django-oauth-toolkit==2.3.0
django-osm-field==0.5.0
djangorestframework==3.15.1
drf-yasg==1.21.7
ez_setup==0.9
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1

geopy~=2.4.1